made putting documents into firebase more OOP
