+++ src/firebase/putRecipe.js
@@ -1,40 +1,11 @@
 import { doc, getDoc, setDoc, getFirestore, toFirestore, fromFirestore } from "firebase/firestore";
 import { firebaseApp } from "./firebaseConfig.js";
+import FirebaseConverter from "../managers_and_parsers/FirebaseConverter.js"
 
 const db = getFirestore(firebaseApp);
 
-const recipeConverter = {
-
-    toFirestore: (recipe) => {
-        return {
-            cuisine: recipe.cuisine,
-            dishType: recipe.dishType,
-            id: recipe.id,
-            image: recipe.image,
-            ingredients: recipe.ingredients,
-            instructions: recipe.instructions,
-            name: recipe.name,
-            servings: recipe.servings,
-            summary: recipe.summary
-        };
-    },
-
-    fromFirestore: (snapshot, options) => {
-        const data = snapshot.data(options);
-        return new Recipe(    
-            data.cuisine,
-            data.dishType,
-            data.id,
-            data.image,
-            data.ingredients,
-            data.instructions,
-            data.name,
-            data.servings,
-            data.summary,
-            )
-        }
-    }
-
+const fb = new FirebaseConverter()
+const recipeConverter = fb.recipeConverter;
 
 /**
  * @param {String} collection the collection to save into
